+++ CHANGELOG.md
@@ -9,8 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
-- Add certificate issuer label to `cert_exporter_secret_not_after` and `cert_exporter_not_after` metrics.
-- Add certificate CR name to `cert_exporter_secret_not_after` metric.
+- Add new `cert_exporter_certificate_cr_not_after` metric. This metric exports the `spec.notAfter` field of cert-manager `Certificate` CR.
 
 ### Changed