+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [1.8.0] - 2021-08-25
+
 ### Added
 
 - Add new `cert_exporter_certificate_cr_not_after` metric. This metric exports the `status.notAfter` field of cert-manager `Certificate` CR.
@@ -100,7 +102,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 - Tolerations changed to tolerate all taints.
 - Change priority class to `giantswarm-critical`.
 
-[Unreleased]: https://github.com/giantswarm/cert-exporter/compare/v1.7.1...HEAD
+[Unreleased]: https://github.com/giantswarm/cert-exporter/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/giantswarm/cert-exporter/compare/v1.7.1...v1.8.0
 [1.7.1]: https://github.com/giantswarm/cert-exporter/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/giantswarm/cert-exporter/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/giantswarm/cert-exporter/compare/v1.6.0...v1.6.1